Modificar Controlador catFeriados
Implementación de auditoria en los métodos del controlador
Implementación de transacciones

+++ servidor/src/controlador/catalogos/catFeriadosControlador.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
 import { QueryResult } from 'pg';
 import { ObtenerRutaLeerPlantillas } from '../../libs/accesoCarpetas';
+import AUDITORIA_CONTROLADOR from '../auditoria/auditoriaControlador';
 import moment from 'moment';
 import excel from 'xlsx';
 import pool from '../../database';
@@ -26,20 +27,70 @@ class FeriadosControlador {
     }
 
     // METODO PARA ELIMINAR UN REGISTRO DE FERIADOS
-    public async EliminarFeriado(req: Request, res: Response): Promise<any> {
-        const id = req.params.id;
-        await pool.query(
-            `
-            DELETE FROM cg_feriados WHERE id = $1
-            `
-            , [id]);
-        res.jsonp({ text: 'Registro eliminado.' });
+    public async EliminarFeriado(req: Request, res: Response): Promise<Response> {
+        try {
+            const id = req.params.id;
+            // TODO ANALIZAR COMO OBTENER DESDE EL FRONT EL USERNAME Y LA IP
+
+            const {user_name, ip} = req.body;
+
+            // INICIAR TRANSACCION
+            await pool.query('BEGIN');
+
+            // CONSULTAR DATOS ORIGINALES
+
+            const datosOriginales = await pool.query('SELECT * FROM cg_feriados WHERE id = $1', [id]);
+            const [feriado] = datosOriginales.rows;
+
+            if (!feriado) {
+                await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                    tabla: 'cg_feriados',
+                    usuario: user_name,
+                    accion: 'D',
+                    datosOriginales: '',
+                    datosNuevos: '',
+                    ip,
+                    observacion: `Error al eliminar feriado con id ${id}.`
+                })
+
+                // FINALIZAR TRANSACCION
+                await pool.query('COMMIT');
+                return res.status(404).jsonp({ text: 'Registro no encontrado.' });
+            }
+            await pool.query(
+                `
+                DELETE FROM cg_feriados WHERE id = $1
+                `
+                , [id]);
+            
+            // AUDITORIA
+            await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                tabla: 'cg_feriados',
+                usuario: user_name,
+                accion: 'D',
+                datosOriginales: JSON.stringify(feriado),
+                datosNuevos: '',
+                ip,
+                observacion: null
+            });
+
+            // FINALIZAR TRANSACCION
+            await pool.query('COMMIT');
+            return res.jsonp({ text: 'Registro eliminado.' });
+        } catch (error) {
+            // REVERTIR TRANSACCION
+            await pool.query('ROLLBACK');
+            return res.jsonp({ message: 'error' });
+        }
     }
 
     // METODO PARA CREAR REGISTRO DE FERIADO
     public async CrearFeriados(req: Request, res: Response): Promise<Response> {
         try {
-            const { fecha, descripcion, fec_recuperacion } = req.body;
+            const { fecha, descripcion, fec_recuperacion, user_name, ip } = req.body;
+
+            // INICIAR TRANSACCION
+            await pool.query('BEGIN');
 
             const response: QueryResult = await pool.query(
                 `
@@ -50,6 +101,20 @@ class FeriadosControlador {
 
             const [feriado] = response.rows;
 
+            // AUDITORIA
+            await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                tabla: 'cg_feriados',
+                usuario: user_name,
+                accion: 'I',
+                datosOriginales: '',
+                datosNuevos: JSON.stringify(feriado),
+                ip,
+                observacion: null
+            });
+
+            // FINALIZAR TRANSACCION
+            await pool.query('COMMIT');
+
             if (feriado) {
                 return res.status(200).jsonp(feriado)
             }
@@ -58,6 +123,8 @@ class FeriadosControlador {
             }
         }
         catch (error) {
+            // REVERTIR TRANSACCION
+            await pool.query('ROLLBACK');
             return res.jsonp({ message: 'error' });
         }
     }
@@ -79,18 +146,58 @@ class FeriadosControlador {
     }
 
     // METODO PARA ACTUALIZAR UN FERIADO
-    public async ActualizarFeriado(req: Request, res: Response) {
+    public async ActualizarFeriado(req: Request, res: Response): Promise<Response> {
         try {
-            const { fecha, descripcion, fec_recuperacion, id } = req.body;
+            const { fecha, descripcion, fec_recuperacion, id, user_name, ip } = req.body;
+
+            // INICIAR TRANSACCION
+            await pool.query('BEGIN');
+
+            // CONSULTAR DATOS ORIGINALES
+            const datosOriginales = await pool.query('SELECT * FROM cg_feriados WHERE id = $1', [id]);
+            const [feriado] = datosOriginales.rows;
+
+            if (!feriado) {
+                await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                    tabla: 'cg_feriados',
+                    usuario: user_name,
+                    accion: 'U',
+                    datosOriginales: '',
+                    datosNuevos: '',
+                    ip,
+                    observacion: `Error al actualizar feriado con id ${id}.`
+                })
+
+                // FINALIZAR TRANSACCION
+                await pool.query('COMMIT');
+                return res.status(404).jsonp({ text: 'Registro no encontrado.' });
+            }
+
             await pool.query(
                 `
                 UPDATE cg_feriados SET fecha = $1, descripcion = $2, fec_recuperacion = $3
                 WHERE id = $4
                 `
                 , [fecha, descripcion, fec_recuperacion, id]);
-            res.jsonp({ message: 'Registro actualizado.' });
+            
+            // AUDITORIA
+            await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                tabla: 'cg_feriados',
+                usuario: user_name,
+                accion: 'U',
+                datosOriginales: JSON.stringify(feriado),
+                datosNuevos: JSON.stringify({ fecha, descripcion, fec_recuperacion }),
+                ip,
+                observacion: null
+            });
+
+            // FINALIZAR TRANSACCION
+            await pool.query('COMMIT');
+            return res.jsonp({ message: 'Registro actualizado.' });
         }
         catch (error) {
+            // REVERTIR TRANSACCION
+            await pool.query('ROLLBACK');
             return res.jsonp({ message: 'error' });
         }
     }
@@ -159,16 +266,6 @@ class FeriadosControlador {
             return res.jsonp({ message: 'error' });
         }
     }
-
-
-
-
-
-
-
-
-
-
 
     /* 
     * Metodo para revisar
@@ -482,6 +579,7 @@ class FeriadosControlador {
     // INGRESAR DATOS DE FERIADOS MEDIANTE PLANTILLA
     public async CrearFeriadoPlantilla(req: Request, res: Response): Promise<void> {
         const documento = req.file?.originalname;
+        const { user_name, ip } = req.body;
         let separador = path.sep;
         let ruta = ObtenerRutaLeerPlantillas() + separador + documento;
         var contador = 1;
@@ -490,23 +588,46 @@ class FeriadosControlador {
         const plantilla = excel.utils.sheet_to_json(workbook.Sheets[sheet_name_list[2]]);
         // LECTURA DE DATOS DE LA PLANTILLA
         plantilla.forEach(async (data: any) => {
-            const { fecha, descripcion, fec_recuperacion } = data;
-            // VERIFICACIÓN DE EXISTENCIA DE DATOS DE FECHA DE RECUPERACIÓN
-            if (fec_recuperacion === undefined) {
-                var recuperar = null;
-            }
-            else {
-                recuperar = fec_recuperacion;
-            }
-            // REGISTRO DE DATOS EN EL SISTEMA
-            await pool.query('INSERT INTO cg_feriados (fecha, descripcion, fec_recuperacion) ' +
-                'VALUES ($1, $2, $3)', [fecha, descripcion, recuperar]);
+            try {
+                const { fecha, descripcion, fec_recuperacion } = data;
+                // VERIFICACIÓN DE EXISTENCIA DE DATOS DE FECHA DE RECUPERACIÓN
+                if (fec_recuperacion === undefined) {
+                    var recuperar = null;
+                }
+                else {
+                    recuperar = fec_recuperacion;
+                }
 
-            // ENVIO DE MENSAJE UNA VEZ QUE SE HA LEIDO TODOS LOS DATOS DE LA PLANTILLA
-            if (contador === plantilla.length) {
-                return res.jsonp({ message: 'CORRECTO' });
+                // INICIAR TRANSACCION
+                await pool.query('BEGIN');
+
+                // REGISTRO DE DATOS EN EL SISTEMA
+                await pool.query('INSERT INTO cg_feriados (fecha, descripcion, fec_recuperacion) ' +
+                    'VALUES ($1, $2, $3)', [fecha, descripcion, recuperar]);
+
+                // AUDITORIA
+                await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                    tabla: 'cg_feriados',
+                    usuario: user_name,
+                    accion: 'I',
+                    datosOriginales: '',
+                    datosNuevos: JSON.stringify({ fecha, descripcion, recuperar }),
+                    ip,
+                    observacion: null
+                });
+
+                // FINALIZAR TRANSACCION
+                await pool.query('COMMIT');
+    
+                // ENVIO DE MENSAJE UNA VEZ QUE SE HA LEIDO TODOS LOS DATOS DE LA PLANTILLA
+                if (contador === plantilla.length) {
+                    return res.jsonp({ message: 'CORRECTO' });
+                }
+                contador = contador + 1;
+            } catch (error) {
+                // REVERTIR TRANSACCION
+                await pool.query('ROLLBACK');
             }
-            contador = contador + 1;
         });
 
         // VERIFICAR EXISTENCIA DE CARPETA O ARCHIVO
@@ -518,9 +639,6 @@ class FeriadosControlador {
             }
         });
     }
-
-
-
 
 }
